Split MCP runtime tool tests
